Make logger take a custom low level `writer` function to allow customization of interface to Logentries.
Summary: Logentries provides a REST api for logging, but no official client. We want to log this way, because it will work across a web proxy and from behind a firewall. Almost all of the changes, save for some very minor ones, involve ripping out functionality related to maintaining the health of the (former) TCP connection, or exposing previously-private internal details for customization. In particular, the user must provide a `writer` function, which does actual communication with Logentries server. For this particular use case, it simply does an HTTP post.

Test Plan: Tested manually by integrating with Player. It should be a drop in replacement for mainline `le_node` **except** for the the need to provide the `writer` function.

Reviewers: micahrye

Reviewed By: micahrye

Differential Revision: https://phab.tulip.co/D3173