fix bug when `writer`'s returned promise resolves to a value
Summary: The _write callback was being chained directly to the promise returned by the client-provided `writer`. If that resolved to a value, it was passed through to the callback, which caused it to bug out. This fix makes sure that the callback is always called with no args.

Test Plan: Manually by integrating w/player

Reviewers: micahrye

Reviewed By: micahrye

Differential Revision: https://phab.tulip.co/D3199